Moving JS files to src/js and adding some logging output for the build process.
