+++ src/js/pagedata.js
@@ -2,6 +2,7 @@ const glob = require('glob');
 const fse = require('fs-extra');
 const frontMatter = require('front-matter');
 const path = require('path');
+const logger = require('./logger');
 
 const getLink = (file) => {
     fileData = path.parse(file);
@@ -12,6 +13,7 @@ const getPageData = (srcDirectory, layout) => {
 
     var pages = [];
   
+    logger.info(`Searching: ${srcDirectory} (for layout type: ${layout})`);
     const files = glob.sync('**/*.@(md|ejs|html)', { cwd: `${srcDirectory}` });
 
     files.forEach((file) => {
@@ -29,6 +31,7 @@ const getPageData = (srcDirectory, layout) => {
         page.link = getLink(file);
       }
   
+      logger.info(` + ${page.link}`);    
       pages.push(page);
     });
